% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bp_location_posterior}
\alias{bp_location_posterior}
\title{Creates plot of posterior pulse locations}
\usage{
bp_location_posterior(fit, ...)
}
\arguments{
\item{fit}{A \code{pulse_fit} or \code{pop_pulse_fit} object}

\item{...}{Additional arguments. Use \code{patient} argument to determine
which patient's trace plots are shown when \code{fit} is a
\code{pop_pulse_fit} object.}
}
\description{
Creates plot of posterior pulse locations
}
