% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_conc.R
\name{bp_predicted.pop_pulse_fit}
\alias{bp_predicted.pop_pulse_fit}
\title{Plot predicted concentration, credible interval, and observed data}
\usage{
\method{bp_predicted}{pop_pulse_fit}(fit, predicted)
}
\arguments{
\item{fit}{a pop_fit_pulse object}

\item{predicted}{result of predict.pulse_fit(fit)}
}
\description{
Plot predicted concentration, credible interval, and observed data
}
