% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bp_predicted}
\alias{bp_predicted}
\title{Plot predicted concentration, credible interval, and observed data}
\usage{
bp_predicted(fit, predicted)
}
\arguments{
\item{fit}{A \code{pulse_fit} or \code{pop_pulse_fit} object}

\item{predicted}{Result of \code{predict()} run on result of 
predict.pulse_fit(fit)}
}
\description{
Plot predicted concentration, credible interval, and observed data
}
\examples{

this_pulse <- simulate_pulse()
this_spec  <- pulse_spec(location_prior_type = "strauss",
                         prior_location_gamma = 0,
                         prior_location_range = 40)
this_fit   <- fit_pulse(data = this_pulse, iters = 1000, thin = 10,
                        spec = this_spec)
fit_predicted <- predict(this_fit, cred_interval = 0.8)
bp_predicted(this_fit, fit_predicted)

}
\keyword{diagnostics}
\keyword{fit}
\keyword{plot}
\keyword{predicted}
\keyword{pulse}
