% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diagnostics.R
\name{bp_trace}
\alias{bp_trace}
\alias{bp_posteriors.pulse_fit}
\title{Creates chain trace plots}
\usage{
bp_trace(fit, ...)

\method{bp_posteriors}{pulse_fit}(fit, type = c("histogram", "density"))
}
\description{
Creates chain trace plots
}
