% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{erf}
\alias{erf}
\title{Creates the integral in the deconvolution part of the model}
\usage{
erf(time)
}
\arguments{
\item{time}{Some time variable}
}
\description{
Creates the integral in the deconvolution part of the model
}
