% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pop.R
\name{fit_pop_pulse}
\alias{fit_pop_pulse}
\title{fit_pop_pulse}
\usage{
fit_pop_pulse(
  data,
  time,
  num_patients,
  location_prior = "strauss",
  spec,
  iters = 250000,
  thin = 50,
  burnin = as.integer(0.1 * iters),
  use_tibble = TRUE,
  verbose = FALSE,
  test_birthdeath = FALSE,
  test_sd_masses = FALSE,
  test_sd_width = FALSE,
  test_s2s_sd_width = FALSE,
  test_s2s_sd_mass = FALSE,
  test_s2s_sd_baseline = FALSE,
  test_s2s_sd_halflife = FALSE,
  test_pop_means_width = FALSE,
  test_pop_means_mass = FALSE,
  test_pop_means_baseline = FALSE,
  test_pop_means_halflife = FALSE,
  test_fixeff_mass = FALSE,
  test_fixeff_width = FALSE,
  test_error = FALSE,
  test_locations = FALSE,
  test_masses = FALSE,
  test_widths = FALSE
)
}
\arguments{
\item{data}{Placeholder}

\item{time}{A string. Name of the time variable in \code{data}}

\item{num_patients}{Number of patients in the analysis}

\item{spec}{An object of class \code{spec}, created by
\code{pulse_spec()}, specifying the priors, starting values, and proposal
variances to use.}

\item{iters}{Number of iterations for MCMC}

\item{thin}{Thinning to apply to MCMC chains (i.e. Keep every 'thin'th
sample).}

\item{burnin}{Burn-in to apply to MCMC chains (i.e. remove first 'burnin'
samples). Applied prior to thinning.}

\item{use_tibble}{Return chains as tbl_df class data frames, from the tibble
package.  Mostly used for the print.tbl_df method, which limits the rows and
columns printed to those which fit in the console.}

\item{verbose}{Prints diagnostics and estimates every 5000th iteration.
Default is \code{FALSE}.}
}
\description{
fit_pop_pulse
}
