% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_pulse}
\alias{fit_pulse}
\title{fit_pulse}
\usage{
fit_pulse(
  data,
  time = "time",
  conc = "concentration",
  subject_id = NULL,
  spec,
  iters = 250000,
  thin = 50,
  burnin = as.integer(0.1 * iters),
  use_tibble = TRUE,
  verbose = FALSE,
  verbose_iter = 5000,
  fix_params = c("width_sd"),
  pulse_masses = numeric(),
  pulse_widths = numeric(),
  pulse_mass_sdscales = numeric(),
  pulse_width_sdscales = numeric(),
  pulse_locations = numeric()
)
}
\arguments{
\item{data}{A dataset containing \code{time} and \code{conc}.  Can
also be a \code{pulse_sim} object.}

\item{time}{A string. Name of the time variable in \code{data}.}

\item{conc}{A string. Name of the hormone concentration variable in
\code{data}.}

\item{subject_id}{A string. Name of the subject id variable in \code{data}.}

\item{spec}{An object of class \code{spec}, created by
\code{pulse_spec()}, specifying the priors, starting values, and proposal
variances to use.}

\item{iters}{Number of iterations for MCMC}

\item{thin}{Thinning to apply to MCMC chains (i.e. Keep every 'thin'th
sample).}

\item{burnin}{Burn-in to apply to MCMC chains (i.e. remove first 'burnin'
samples). Applied prior to thinning.}

\item{use_tibble}{Return chains as tbl_df class data frames, from the tibble
package.  Mostly used for the print.tbl_df method, which limits the rows and
columns printed to those which fit in the console.}

\item{verbose}{Prints diagnostics and estimates every 5000th iteration.
Default is \code{FALSE}.}

\item{verbose_iter}{Integer that sets diagnostic output to printout at the
given interval if 'verbose' is \code{TRUE}. Default is 5000.}

\item{fix_params}{Character vector enabling parameters to fixed (i.e. not 
estimated). Vector may include options:
"pulse_count", "mass_mean", "width_mean", "mass_sd", 
"width_sd", "bl_hl", "pulse_location", "error_var",
"pulse_mass", "pulse_width", "pulse_mass_sdscale", 
"pulse_width_sdscale" 
or may be left empty to estimate all parameters. For patient level 
parameters, values are fixed at the starting values specified by
\code{pulse_spec()}. For pulse level parameters, vectors must be
provided using the function arguments below to specify their fixed values.}

\item{pulse_masses}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse masses. 
If used, \code{pulse_count} must also be fixed, and the length of the 
vector must equal the sum of \code{pulse_counts}.}

\item{pulse_widths}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse widths. If
used, \code{pulse_count} must also be fixed, and length of the vector 
must equal the sum of \code{prior_mean_pulse_count}.}

\item{pulse_mass_sdscales}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse mass standard
deviation scales. If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{prior_mean_pulse_count}.}

\item{pulse_width_sdscales}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse width standard
deviation scales. If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{prior_mean_pulse_count}.}

\item{pulse_locations}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse locations. 
If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{prior_mean_pulse_count}.}
}
\description{
Primary function for fitting deconvolution model for time series of pulsatile
hormone data.  Expects a tidy, long-format dataset (columns for time,
concentration, and subject ID; row for every concentration measurement).
}
\examples{
this_pulse <- simulate_pulse()
this_spec  <- pulse_spec()
#this_fit   <- fit_pulse(data = this_pulse, iters = 1000, thin = 10, 
#                        burnin = 100, spec = this_spec)
}
\keyword{fit}
\keyword{pulse}
