% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{patient_chain}
\alias{patient_chain}
\title{Patient level MCMC chain for a patient}
\usage{
patient_chain(fit, ...)
}
\arguments{
\item{fit}{A \code{pulse_fit} or \code{pop_pulse_fit} object}

\item{...}{Additional arguments. Use \code{patient} argument to determine
which patient's MCMC chain is returned from a \code{pop_pulse_fit}
object}
}
\description{
Patient level MCMC chain for a patient
}
