% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pop.R
\name{plot.pop_sim}
\alias{plot.pop_sim}
\title{Plot a simulated population of pulsatile hormone time series.}
\usage{
\method{plot}{pop_sim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pulse_sim} resulting from the
\code{simulate_pulse()} function.}

\item{...}{Other arguments not used by this method.}
}
\value{
A ggplot2 plot of the \code{simulate_pulse} time-series dataset.
}
\description{
\code{\link{plot.pop_sim}} plots the time-series components of
a \code{pop_sim} object
}
\examples{
this_pop <- simulate_pop()
plot(this_pop)
}
\seealso{
simulate_pop, print.pop_sim, summary.pop_sim
}
\keyword{pulse}
\keyword{simulation}
