% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot.pulse_sim}
\alias{plot.pulse_sim}
\title{Plot a simulated pulsatile hormone time series.}
\usage{
\method{plot}{pulse_sim}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pulse_sim} resulting from the
\code{simulate_pulse()} function.}

\item{...}{Other arguments not used by this method.}
}
\value{
A ggplot2 plot of the \code{simulate_pulse} time-series dataset.
}
\description{
\code{\link{plot.pulse_sim}} plots the time-series component of
a \code{pulse_sim} object
}
\examples{
this_pulse <- simulate_pulse()
plot(this_pulse)
}
\seealso{
simulate_pulse, print.pulse_sim, summary.pulse_sim
}
\keyword{pulse}
\keyword{simulation}
