% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot Autocorrelation Functions}
\usage{
plot_acf(fit, parameters = NULL, max_lag = 50, ncol = 2)
}
\arguments{
\item{fit}{A \code{population_fit} object}

\item{parameters}{Character vector of parameter names to plot. If NULL,
plots all population parameters.}

\item{max_lag}{Maximum lag to display (default: 50)}

\item{ncol}{Number of columns in plot grid (default: 2)}
}
\value{
A ggplot2 object
}
\description{
Creates autocorrelation plots for population-level parameters.
}
