% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsespec_pop.R
\name{pop_spec}
\alias{pop_spec}
\title{pop_spec}
\usage{
pop_spec(
  prior_mass_mean,
  prior_mass_var,
  prior_mass_p2p_sd_var_max,
  prior_mass_s2s_sd_var_max,
  prior_width_mean,
  prior_width_var,
  prior_width_p2p_sd_var_max,
  prior_width_s2s_sd_var_max,
  prior_baseline_mean,
  prior_baseline_var,
  prior_baseline_s2s_sd_var_max,
  prior_halflife_mean,
  prior_halflife_var,
  prior_halflife_s2s_sd_var_max,
  sv_mass_mean,
  sv_mass_p2p_sd_var,
  sv_mass_s2s_sd_var,
  sv_width_mean,
  sv_width_p2p_sd_var,
  sv_width_s2s_sd_var,
  sv_baseline_mean,
  sv_baseline_s2s_sd_var,
  sv_halflife_mean,
  sv_halflife_s2s_sd_var,
  sv_error_var,
  sv_error_alpha,
  sv_error_beta,
  sv_mean_pulse_count,
  sv_strauss_repulsion,
  sv_strauss_repulsion_range,
  pv_sub_mass_mean,
  pv_pop_mass_mean,
  pv_ind_mass,
  pv_sub_width_mean,
  pv_pop_width_mean,
  pv_ind_width,
  pv_pop_baseline_mean,
  pv_sub_baseline,
  pv_pop_halflife_mean,
  pv_sub_halflife,
  pv_pulse_location
)
}
\arguments{
\item{prior_mass_mean}{Placeholder}

\item{prior_mass_variance}{Placeholder}
}
\description{
pop_spec
}
