% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppulsatile-package.R
\docType{package}
\name{poppulsatile-package}
\alias{poppulsatile-package}
\alias{poppulsatile}
\title{poppulsatile package documentation}
\description{
Bayesian deconvolution model for the analysis of time-series of
pulsatile hormone data. See the help files for the primary functions
pulse_spec() and fit_pulse().
}
