% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{population_chain}
\alias{population_chain}
\title{Population level MCMC chain}
\usage{
population_chain(fit)
}
\arguments{
\item{fit}{A  \code{pop_pulse_fit} object}
}
\description{
Population level MCMC chain
}
