% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.pop_pulse_fit}
\alias{predict.pop_pulse_fit}
\title{Predict concentrations at each observation time}
\usage{
\method{predict}{pop_pulse_fit}(object, ...)
}
\arguments{
\item{object}{A model fit from \code{fit_pulse()} (class "pulse_fit") or
from \code{fit_pop_pulse()} (class "pop_pulse_fit").}

\item{...}{Additional arguments affecting the predictions produced. Use
\code{cred_interval} to alter size of credible interval calculated.}
}
\description{
Calculates the predicted concentration and specified credible interval for 
each observation time.
}
