% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_conc.R
\name{predict.pulse_fit}
\alias{predict.pulse_fit}
\alias{calc_mean_contrib}
\alias{add_baseline_error}
\title{Predict concentrations at each observation time}
\usage{
\method{predict}{pulse_fit}(object, ...)

calc_mean_contrib(time, location, halflife, mass, width)

add_baseline_error(x, baseline, model_error)
}
\arguments{
\item{object}{A model fit from \code{fit_pulse()} (class "pulse_fit") or
from \code{fit_pop_pulse()} (class "pop_pulse_fit").}

\item{...}{Additional arguments affecting the predictions produced. Use
\code{cred_interval} to alter size of credible interval calculated.}

\item{time}{Vector of times from fitstart to fitend, by minute? sampling
interval?}

\item{location}{Scalar. Location of this pulse}

\item{halflife}{Scalar. Halflife for this iteration.}

\item{mass}{Scalar. Mass of this pulse.}

\item{width}{Scalar. Width of this pulse}

\item{x}{Mean concentration (elementwise sum of mean_contribution vectors).}

\item{baseline}{Baseline concentration value for this iteration.}

\item{model_error}{Model error estimate for this iteration.}
}
\description{
Calculates the predicted concentration and specified credible interval for 
each observation time.
}
\examples{

this_pulse <- simulate_pulse()
this_spec  <- pulse_spec(location_prior_type = "strauss",
                         prior_location_gamma = 0,
                         prior_location_range = 40)
this_fit   <- fit_pulse(data = this_pulse, iters = 1000, thin = 10,
                        spec = this_spec)
fit_predicted <- predict(this_fit, cred_interval = 0.8)

}
\keyword{diagnostics}
\keyword{fit}
\keyword{plot}
\keyword{predicted}
\keyword{pulse}
