% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsatile-package.R
\name{pulsatile}
\alias{pulsatile}
\alias{pulsatile-package}
\title{Pulsatile package documentation}
\description{
Bayesian deconvolution model for the analysis of time-series of
pulsatile hormone data. See the help files for the primary functions
pulse_spec() and fit_pulse().
}
