% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pulse_chain}
\alias{pulse_chain}
\title{Pulse level MCMC chain for a patient}
\usage{
pulse_chain(fit, ...)
}
\arguments{
\item{fit}{A \code{pulse_fit} or \code{pop_pulse_fit} object}

\item{...}{Additional arguments. Use \code{patient} argument to determine
which patient's pulse chain is returned from a \code{pop_pulse_fit}
object}
}
\description{
Pulse level MCMC chain for a patient
}
