% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pop.R
\name{simulate_pop}
\alias{simulate_pop}
\title{Simulate pulsatile hormone data for a population of patients}
\usage{
simulate_pop(
  num_pats = 10,
  num_obs = 144,
  interval = 10,
  error_var = 0.005,
  ipi_mean = 12,
  ipi_var = 40,
  ipi_min = 4,
  pop_mass_mean = 3.5,
  mass_s2s_sd = 1,
  mass_p2p_sd = 1.6,
  pop_width_mean = 35,
  width_s2s_sd = 1,
  width_p2p_sd = 5,
  pop_halflife_mean = 45,
  halflife_sd = 5,
  pop_baseline_mean = 2.5,
  baseline_sd = 0.75
)
}
\arguments{
\item{num_pats}{Number of patients to simulate.}

\item{num_obs}{Number of observations to simulate per patient. Duration of 
observation window equals \code{num_obs} times \code{interval}.}

\item{interval}{Time in minutes between observations, typically 6-10.}

\item{error_var}{Variance of the error added at each observation, error ~ N(0, sqrt(error_var)).}

\item{ipi_mean}{Mean number of sampling units between pulses (mean inter-pulse interval).}

\item{ipi_var}{Variance of gamma for drawing interpulse interval}

\item{ipi_min}{Minimum number of units between pulses}

\item{pop_mass_mean}{Population pulse mass}

\item{mass_s2s_sd}{Standard deviation of patient pulse mass means}

\item{mass_p2p_sd}{Standard deviation of pulse masses}

\item{pop_width_mean}{Population mean pulse width (in minutes)}

\item{width_s2s_sd}{Standard deviation of patient pulse width means (in minutes)}

\item{width_p2p_sd}{Standard deviation of pulse width (in minutes)}

\item{pop_halflife_mean}{Population mean of half-life (in minutes)}

\item{halflife_sd}{Variance of patient half-lives (in minutes)}

\item{pop_baseline_mean}{Population mean of baseline}

\item{baseline_sd}{Variance of patient baselines}
}
\value{
A object of class \code{pop_sim} containing time-series data set
  and data set of characteristics of each pulse
}
\description{
\code{\link{simulate_pop}} simulates multiple time series data sets
  representing blood concentration measurements of a pulsatile hormone. A
  set of time series as well as \code{data.frame}'s of population, patient,
  and individual pulse characteristics are returned.
}
\examples{
this_pop <- simulate_pop()
str(this_pop)
plot(this_pop)
}
\seealso{
print.pulse_sim, plot.pulse_sim
}
\keyword{pulse}
\keyword{simulation}
