% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_joint.R
\name{summary.joint_fit}
\alias{summary.joint_fit}
\title{Summary method for joint_fit objects}
\usage{
\method{summary}{joint_fit}(object, ...)
}
\arguments{
\item{object}{A joint_fit object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for joint_fit objects
}
