% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayespulse-package.R
\docType{package}
\name{bayespulse-package}
\alias{bayespulse-package}
\alias{bayespulse}
\title{bayespulse package documentation}
\description{
Bayesian deconvolution model for the analysis of time-series of
pulsatile hormone data. See the help files for the primary functions
pulse_spec() and fit_pulse().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BayesPulse/bayespulse}
  \item Report bugs at \url{https://github.com/BayesPulse/bayespulse/issues}
}

}
\author{
\strong{Maintainer}: Matt Mulvahill \email{matthew.mulvahill@ucdenver.edu}

Authors:
\itemize{
  \item Nichole Carlson \email{nichole.carlson@ucdenver.edu} [copyright holder, thesis advisor]
}

}
