% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bp_location_posterior}
\alias{bp_location_posterior}
\title{Creates plot of posterior pulse locations}
\usage{
bp_location_posterior(fit, ...)
}
\description{
Creates plot of posterior pulse locations
}
