% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{bp_trace}
\alias{bp_trace}
\alias{bp_posteriors}
\alias{bp_location_posterior}
\title{Creates chain trace plots}
\usage{
bp_trace(fit, ...)

bp_posteriors(fit, type = c("histogram", "density"))

bp_location_posterior(fit)
}
\description{
Creates chain trace plots
}
