% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{bp_trace}
\alias{bp_trace}
\title{Creates chain trace plots}
\usage{
bp_trace(fit, ...)
}
\arguments{
\item{fit}{A \code{pulse_fit} or \code{pop_pulse_fit} object}

\item{...}{Additional arguments. Use \code{patient} argument to determine
which patient's trace plots are shown when \code{fit} is a
\code{pop_pulse_fit} object.}
}
\description{
Creates chain trace plots
}
\examples{
this_pulse <- simulate_pulse()
this_spec  <- pulse_spec()
this_fit   <- fit_pulse(data = this_pulse, iters = 1000, thin = 10,
                        spec = this_spec)
bp_trace(this_fit)
}
\keyword{diagnostics}
\keyword{fit}
\keyword{plot}
\keyword{pulse}
