% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{bp_trace}
\alias{bp_trace}
\alias{bp_posteriors}
\alias{bp_location_posterior}
\title{Diagnostic plots for \code{fit_pulse()} models}
\usage{
bp_trace(fit)

bp_posteriors(fit, type = c("histogram", "density"))

bp_location_posterior(fit)
}
\arguments{
\item{fit}{A model fit from \code{fit_pulse()}.}

\item{type}{Either histogram or density.  Only applies to
\code{bp_posteriors} function}
}
\description{
Plotting functions for mcmc chains from \code{fit_pulse()} models.  Includes
trace plots and posterior densities of the 'common' parameters and pulse
location density (a set of pulse-specific parameter, from 'pulse' chain).
}
\examples{
this_pulse <- simulate_pulse()
this_spec  <- pulse_spec()
this_fit   <- fit_pulse(.data = this_pulse, iters = 1000, thin = 10,
                        spec = this_spec)
bp_trace(this_fit)
}
\keyword{diagnostics}
\keyword{fit}
\keyword{plot}
\keyword{pulse}
