% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{chains}
\alias{chains}
\alias{population_chain}
\alias{pulse_chain}
\alias{patient_chain}
\title{chains}
\usage{
chains(fit)

population_chain(fit)

pulse_chain(fit, ...)

patient_chain(fit, ...)
}
\arguments{
\item{fit}{A model fit from \code{fit_pulse}.}
}
\description{
Functions for extracting and working with the various levels of posterior
mcmc chains.
}
\examples{

pulse <- simulate_pulse()
spec  <- pulse_spec()
fit   <- fit_pulse(data = pulse, iters = 1000, thin = 10,
                   burnin = 100, spec = spec)
#chains(fit)
#pulse_chain(fit)
#patient_chain(fit)

}
\keyword{fit}
\keyword{pulse}
