% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pop.R
\name{fit_pop_pulse}
\alias{fit_pop_pulse}
\title{fit_pop_pulse}
\usage{
fit_pop_pulse(
  data,
  time,
  location_prior = "strauss",
  spec,
  iters = 250000,
  thin = 50,
  burnin = as.integer(0.1 * iters),
  use_tibble = TRUE,
  verbose = FALSE,
  verbose_patient = FALSE,
  verbose_iter = 5000,
  fix_params = NULL,
  pat_mass_means = numeric(),
  pat_width_means = numeric(),
  pat_baselines = numeric(),
  pat_halflives = numeric(),
  pulse_counts = numeric(),
  pulse_masses = numeric(),
  pulse_widths = numeric(),
  pulse_mass_sdscales = numeric(),
  pulse_width_sdscales = numeric(),
  pulse_locations = numeric()
)
}
\arguments{
\item{data}{Numeric matrix specifying pulsatile hormone concentration for each
patient in population at each time of observation, with each patient as a 
column and each row as an observation time. No additional columns should be
present.}

\item{time}{Numeric vector for observation times. Must be the same length as
the number of rows in \code{data}}

\item{spec}{An object of class \code{spec}, created by
\code{pulse_spec()}, specifying the priors, starting values, and proposal
variances to use.}

\item{iters}{Number of iterations for MCMC}

\item{thin}{Thinning to apply to MCMC chains (i.e. Keep every 'thin'th
sample).}

\item{burnin}{Burn-in to apply to MCMC chains (i.e. remove first 'burnin'
samples). Applied prior to thinning.}

\item{use_tibble}{Return chains as tbl_df class data frames, from the tibble
package.  Mostly used for the print.tbl_df method, which limits the rows and
columns printed to those which fit in the console.}

\item{verbose}{Prints diagnostics and estimates at iterations determined by
'verbose_iters' argument. Default is \code{FALSE}.}

\item{verbose_patient}{Expands diagnostic output to include patient level
parameters. Default is \code{FALSE}.}

\item{fix_params}{Character vector enabling parameters to fixed (i.e. not 
estimated). Vector may include options:
"pop_mass_mean", "pop_width_mean", "pop_baseline_mean",
"pop_halflife_mean", "mass_s2s_sd", "width_s2s_sd", "baseline_s2s_sd",
"halflife_s2s_sd", "mass_p2p_sd", "width_p2p_sd", "pat_mass_mean",
"pat_width_mean", "pat_bl_hl", "pat_error", "pulse_count", "pulse_location",
"pulse_mass", "pulse_width", "pulse_mass_sdscale", "pulse_width_sdscale", 
or may be left empty to estimate all parameters. For population level 
parameters, values are fixed at the starting values specified by
\code{pop_spec()}. For patient and pulse parameters, vectors must be
provided using the function arguments below to specify their fixed values.}

\item{pat_mass_means}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for patient mass means. Must
be the same length as the number of patients being analyzed (the number
of columns passed to \code{data}).}

\item{pat_width_means}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for patient width means. Must
be the same length as the number of patients being analyzed (the number
of columns passed to \code{data}).}

\item{pat_baselines}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for patient baseline means. Must
be the same length as the number of patients being analyzed (the number
of columns passed to \code{data}).}

\item{pat_halflives}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for patient halflife means. Must
be the same length as the number of patients being analyzed (the number
of columns passed to \code{data}).}

\item{pulse_counts}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for patient pulse numbers. Must
be the same length as the number of patients being analyzed (the number
of columns passed to \code{data}).}

\item{pulse_masses}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse masses. 
If used, \code{pulse_count} must also be fixed, and the length of the 
vector must equal the sum of \code{pulse_counts}.}

\item{pulse_widths}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse widths. If
used, \code{pulse_count} must also be fixed, and length of the vector 
must equal the sum of \code{pulse_counts}.}

\item{pulse_locations}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse locations. 
If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{pulse_counts}.}

\item{verbose_iters}{Integer that sets diagnostic output to printout at the
given interval if 'verbose' is \code{TRUE}. Default is 5000.}

\item{pulse_mass_sdscale}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse mass standard
deviation scales. If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{pulse_counts}.}

\item{pulse_width_sdscale}{Numeric vector used in conjunction with 
\code{fix_params} to choose fixed values for individual pulse width standard
deviation scales. If used, \code{pulse_count} must also be fixed, and 
length of the vector must equal the sum of \code{pulse_counts}.}
}
\description{
fit_pop_pulse
}
