% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsespec.R
\name{pulse_spec}
\alias{pulse_spec}
\title{pulse_spec}
\usage{
pulse_spec(location_prior_type = c("order-statistic", "strauss"),
  prior_mass_mean = 4, prior_mass_var = 100, prior_width_mean = 5,
  prior_width_var = 100, prior_baseline_mean = 2.6,
  prior_baseline_var = 100, prior_halflife_mean = 45,
  prior_halflife_var = 100, prior_error_alpha = 1e-04,
  prior_error_beta = 1e-04, prior_location_gamma = NULL,
  prior_location_range = NULL, prior_max_sd_mass = 100,
  prior_max_sd_width = 150, prior_mean_pulse_count = 12, sv_mass_mean = 4,
  sv_width_mean = 5, sv_baseline_mean = 2.6, sv_halflife_mean = 45,
  sv_error_var = 0.25, sv_mass_sd = 2, sv_width_sd = 1,
  pv_baseline = 0.5, pv_halflife = 45, pv_mean_pulse_mass = 2,
  pv_mean_pulse_width = 5, pv_indiv_pulse_mass = 2,
  pv_indiv_pulse_width = 2, pv_sd_pulse_mass = 2, pv_sd_pulse_width = 10,
  pv_sdscale_pulse_mass = 1, pv_sdscale_pulse_width = 1,
  pv_pulse_location = 10)
}
\arguments{
\item{location_prior_type}{Takes on two values: "order-statistic" and
"strauss". "order-statistic" uses every third order statistic of a Uniform
distribution for the pulse location prior and requires specification of the
prior parameter for mean pulse count ("prior_mean_pulse_count").
"strauss" uses the Strauss interacting point-process as a prior and requires
specification of "prior_mean_pulse_count", "prior_location_gamma", and
"prior_location_range".}

\item{prior_mass_mean}{mass mean hyperparm}

\item{prior_mass_var}{mass variance hyperparm}

\item{prior_width_mean}{width mean hyperparm}

\item{prior_width_var}{width variance hyperparm}

\item{prior_baseline_mean}{mean of prior on baseline}

\item{prior_baseline_var}{variance of prior on baseline}

\item{prior_halflife_mean}{mean of prior on half-life}

\item{prior_halflife_var}{variance of prior on half-life}

\item{prior_error_alpha}{placeholder}

\item{prior_error_beta}{placeholder}

\item{prior_location_gamma}{placeholder}

\item{prior_location_range}{placeholder}

\item{prior_max_sd_mass}{placeholder}

\item{prior_max_sd_width}{placeholder}

\item{prior_mean_pulse_count}{placeholder}

\item{sv_mass_mean}{placeholder}

\item{sv_width_mean}{placeholder}

\item{sv_baseline_mean}{placeholder}

\item{sv_halflife_mean}{placeholder}

\item{sv_error_var}{placeholder}

\item{sv_mass_sd}{placeholder}

\item{sv_width_sd}{placeholder}

\item{pv_baseline}{placeholder}

\item{pv_halflife}{placeholder}

\item{pv_mean_pulse_mass}{placeholder}

\item{pv_mean_pulse_width}{placeholder}

\item{pv_indiv_pulse_mass}{placeholder}

\item{pv_indiv_pulse_width}{placeholder}

\item{pv_sd_pulse_mass}{placeholder}

\item{pv_sd_pulse_width}{placeholder}

\item{pv_sdscale_pulse_mass}{placeholder}

\item{pv_sdscale_pulse_width}{placeholder}

\item{pv_pulse_location}{placeholder}
}
\description{
Generates a pulse_spec object -- the specification object required for
fitting a fit_pulse model.
}
\keyword{pulse}
\keyword{simulation}
