% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_pulse}
\alias{simulate_pulse}
\title{Simulate pulsatile hormone data}
\usage{
simulate_pulse(
  num_obs = 144,
  interval = 10,
  error_var = 0.005,
  ipi_mean = 12,
  ipi_var = 40,
  ipi_min = 4,
  mass_mean = 3.5,
  mass_sd = 1.6,
  width_mean = 35,
  width_sd = 5,
  halflife_mean = NULL,
  halflife_var = NULL,
  baseline_mean = NULL,
  baseline_var = NULL,
  constant_halflife = 45,
  constant_baseline = 2.6
)
}
\arguments{
\item{num_obs}{Number of observations to simulate.  Duration of observation
window equals \code{num_obs} times \code{interval}.}

\item{interval}{Time in minutes between observations, typically 6-10.}

\item{error_var}{Variance of the error added at each observation, error ~ N(0, sqrt(error_var)).}

\item{ipi_mean}{Mean number of sampling units between pulses (mean inter-pulse interval).}

\item{ipi_var}{Variance of gamma for drawing interpulse interval}

\item{ipi_min}{Minimum number of units between pulses}

\item{mass_mean}{Mean pulse mass}

\item{mass_sd}{Standard deviation of pulse mass}

\item{width_mean}{Mean pulse width (in minutes)}

\item{width_sd}{Standard deviation of pulse width (in minutes)}

\item{halflife_mean}{Mean of half-life (in minutes)}

\item{halflife_var}{Variance of half-life (in minutes)}

\item{baseline_mean}{Mean of baseline}

\item{baseline_var}{Variance of baseline}

\item{constant_halflife}{To use a constant (specified) half-life, set this
to a constant value [0,inf) in minutes. Mean and variance of half-life are
not used if this is non-null.}

\item{constant_baseline}{To use a constant (specified) baseline, set this to
a constant [0,inf). Mean and variance of baseline are not used if this is
non-null.}
}
\value{
A object of class \code{pulse_sim} containing time-series dataset
  and dataset of characteristics of each pulse
}
\description{
\code{\link{simulate_pulse}} simulates a time series dataset
  representing blood concentration measurements of a pulsatile hormone. Both
  the time series and a dataset of the individual pulse characteristics are
  returned.
}
\examples{
this_pulse <- simulate_pulse()
str(this_pulse)
plot(this_pulse)
}
\seealso{
print.pulse_sim, plot.pulse_sim
}
\keyword{pulse}
\keyword{simulation}
